requirements/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-3-jet==1.0.7
django-ckeditor==5.9.0
django-filter==2.3.0
django-grappelli==2.14.2
django-jet==1.0.8
django-js-asset==1.2.2
django-suit==2.0a1
django-taggit==1.3.0
djangorestframework==3.11.1
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.18.1
idna==2.10
jet==1.0
Markdown==3.2.2
oauth2client==4.1.3
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
